Incorrect decoding of local dates in Safari.
instead of treating `new Date("2018-04-25T11:00")` as a local date, Safari parses it as a UTC date and returns it with the tz difference added. This fix detects the situation by checking TWO local dates (one in winter, one in summer) which should return hour 0, and if one of them is not 0 it's the Safari bug. In that case, and if the date string that is passed is local (no "Z"), we shift it accordingly.

I don't know how to write a node test for this, but the following notebook shows the fix on Safari:
https://observablehq.com/d/7e8b78e5050258ae

Fixes https://github.com/d3/d3-dsv/issues/45