Incorrect decoding of local dates in Safari. (#53)
instead of treating `new Date("2018-04-25T11:00")` as a local date, Safari parses it as a UTC date and returns it with the tz difference added. This fix detects the situation by checking TWO local dates (one in winter, one in summer) which should return hour 0, and if one of them is not 0 it's the Safari bug.

In that case, and if the date string that is passed is local (no "Z"), and has a time ("T"), we switch to the "2019/09/02 11:00" format, which Safari handles correctly.