Only autoType strings
I understand that d3.autoType is somewhat dangerous and only meant to be used in narrow circumstances. But I have still found myself wanting to use it outside the context of d3-dsv, and thus without the guarantee that values will come in as strings. I was happy when `data.map(d3.autoType)` worked (in a case that happened to be all strings), and then surprised when another time it didn't. (I only came across this thanks to @enjalot!)

(Then again, I also actually initially guessed it'd be part of d3-array, and maybe that oughta be warning to me that this is out of scope; maybe if you were writing a more general-purpose autoType, many things would be different.)